requirements.txt
altgraph==0.17
appdirs==1.4.4
backcall==0.2.0
boto3==1.17.91
botocore==1.20.91
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
configparser==5.0.2
cycler==0.10.0
decorator==5.0.9
docker-pycreds==0.4.0
gitdb==4.0.7
GitPython==3.1.18
gql==2.0.0
graphql-core==2.3.2
h5py==3.5.0
idna==3.2
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
line-profiler==3.3.0
matplotlib==3.4.3
matplotlib-inline==0.1.2
mkl-fft==1.3.0
mkl-service==2.3.0
multiprocess==0.70.12.2
ntfy==2.7.0
numpy==1.20.2
olefile==0.46
packaging==21.0
pandas==1.3.2
parso==0.8.2
pathlib==1.0.1
pathos==0.2.8
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.3.1
pox==0.3.0
ppft==1.6.6.4
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
Rx==1.6.1
s3transfer==0.4.2
scikit-learn==1.0
scipy==1.6.3
seaborn==0.11.2
sentry-sdk==1.3.1
shortuuid==1.0.1
slacker==0.14.0
smmap==4.0.0
submitit==1.4.0
subprocess32==3.5.4
tdw==1.8.23.0
tenacity==8.0.1
threadpoolctl==3.0.0
timm==0.4.12
toml==0.10.2
torch==1.9.1+cu111
torchvision==0.9.1
tqdm==4.61.0
traitlets==5.0.5
urllib3==1.26.6
wandb==0.12.1
wcwidth==0.2.5
yacs==0.1.8
zmq==0.0.0
